% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{add_labels}
\usage{
add_labels(parameter_table)
}
\arguments{
\item{parameter_table}{parameter table from lavaan}
}
\description{
Adds labels to all unlabeled parameters of the lavaan model
}
\keyword{internal}
