% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parameter.R
\name{add_parameter}
\alias{add_parameter}
\title{add_parameter}
\usage{
add_parameter(
  id,
  lhs,
  op,
  rhs,
  user = 1,
  block = 1,
  group = 1,
  free,
  ustart = NA,
  exo = 0,
  label,
  plabel,
  start,
  est,
  se = NA
)
}
\description{
Returns data.frame with added parameter
}
\keyword{internal}
