% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mx_model.R
\name{check_mx_model}
\alias{check_mx_model}
\title{check_mx_model}
\usage{
check_mx_model(mx_model, phantom_free)
}
\arguments{
\item{mx_model}{fitted OpenMx model}

\item{phantom_free}{what to free in the phantom variables. Currently only supports "variance"}
}
\description{
Check if the OpenMx model can be translated to bnlearn
}
\keyword{internal}
