% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_to_phantom.R
\name{cov_to_phantom}
\alias{cov_to_phantom}
\title{cov_to_phantom}
\usage{
cov_to_phantom(parameter_table, mx_model, optimize)
}
\arguments{
\item{parameter_table}{parameter table of an OpenMx model (see ?OpenMx::omxLocateParameters)}

\item{mx_model}{fitted OpenMx model}

\item{optimize}{should the substitute model in case of covariances be optimized?}
}
\value{
fitted mx_model with phantom variables
}
\description{
In Bayesian networks, all dependencies between variables must be expressed with
directed "effects". There is no covariance in the same sense as in OpenMx.
Instead, we must replace covariances with effects of a latent phantom variable.
For instance x1 <-> x2 is replaced with ph -> y1; ph -> y2; ph <-> ph. The
resulting model is identical to the initial OpenMx model in terms of fit, but the residual
variance estimates will change. The implementation has been improved by Dr.
Christian Gische
}
\keyword{internal}
