% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bn_model.R
\name{create_bn_model}
\alias{create_bn_model}
\title{create_bn_model}
\usage{
create_bn_model(parameter_table_bn)
}
\arguments{
\item{parameter_table_bn}{parameter table for bn}
}
\value{
bn model string
}
\description{
Creates a model for bnlearn from the bn parameter table
}
\keyword{internal}
